Fixed a merge issue. Allowed server version to be set by the user via the new changes, rather than always the latest version. Client used to raise 2 exceptions if Sherlock did not start. One was for sherlock not starting and another was that it was unable to connect to Sherlock. Now it will raise 1 exception. If Sherlock cannot start it will raise that exception. If Sherlock starts but PySherlock cannot connect it will raise that exception.
