+++ src/ansys/sherlock/core/launcher.py
@@ -84,25 +84,21 @@ def launch_sherlock(
     except Exception as e:
         print(str(e))
         return None
-    server_version = None
+    _server_version = None
     try:
-
-        args = [_get_sherlock_exe_path(year=year, release_number=release_number)]
-        args.append("-grpcPort=" + str(port))
+        sherlock_launch_cmd, _server_version = _get_sherlock_exe_path(
+            year=year, release_number=release_number
+        )
+        args = [sherlock_launch_cmd, "-grpcPort=" + str(port)]
         if single_project_path != "":
             args.append("-singleProject")
             args.append(single_project_path)
         if sherlock_cmd_args != "":
             args.append(f"{shlex.split(sherlock_cmd_args)}")
         print(args)
-        server_version = get_latest_version()
         subprocess.Popen(args)
-    except Exception as e:
-        LOG.error("Error encountered while starting or executing Sherlock, error = %s" + str(e))
 
-    try:
-
-        sherlock = connect_grpc_channel(port, server_version)
+        sherlock = connect_grpc_channel(port, _server_version)
 
         # Check that the gRPC connection is up (timeout after 3 minutes).
         count = 0
@@ -121,7 +117,7 @@ def launch_sherlock(
 
         return sherlock
     except Exception as e:
-        LOG.error(str(e))
+        LOG.error("Error encountered while starting or executing Sherlock, error = %s" + str(e))
 
 
 #  First PySherlock Release "0.2.0"
@@ -144,11 +140,11 @@ def connect_grpc_channel(port=SHERLOCK_DEFAULT_PORT, server_version=None):
     """
     channel_param = f"{LOCALHOST}:{port}"
     channel = grpc.insecure_channel(channel_param)
-    SHERLOCK = Sherlock(channel, server_version)
-    return SHERLOCK
+    sherlock = Sherlock(channel, server_version)
+    return sherlock
 
 
-def _get_base_ansys(year: int = None, release_number: int = None) -> str:
+def _get_base_ansys(year: int = None, release_number: int = None) -> tuple[str, int]:
     supported_installed_versions = {
         env_key: path
         for env_key, path in os.environ.items()
@@ -158,9 +154,11 @@ def _get_base_ansys(year: int = None, release_number: int = None) -> str:
     if year is not None and release_number is not None:
         try:
             year = _extract_sherlock_version_year(year)
-            version_key = f"AWP_ROOT{year}{release_number}"
+
+            sherlock_version = int(f"{year}{release_number}")
+            version_key = f"AWP_ROOT{sherlock_version}"
             if version_key in supported_installed_versions:
-                return supported_installed_versions[version_key]
+                return supported_installed_versions[version_key], sherlock_version
             else:
                 raise ValueError(f"Sherlock {year} {release_number} is not installed.")
         except ValueError as e:
@@ -169,35 +167,29 @@ def _get_base_ansys(year: int = None, release_number: int = None) -> str:
 
     for key in sorted(supported_installed_versions, reverse=True):
         ansys_version = _get_ansys_version_from_awp_root(key)
-
+        sherlock_version = int(ansys_version)
         if ansys_version >= _EARLIEST_SUPPORTED_VERSION:
-            return ansys_version
-    return ""
+            return supported_installed_versions[key], sherlock_version
+
+    raise ValueError("Could not find any installed version of Sherlock.")
 
 
-def _get_base_ansys():
-    supported_installed_versions = {
-        env_key: path
-        for env_key, path in os.environ.items()
-        if env_key.startswith("AWP_ROOT") and os.path.isdir(path)
-    }
+def _get_ansys_version_from_awp_root(awp_root):
+    if awp_root.find("AWP_ROOT") >= 0:
+        return int(awp_root.replace("AWP_ROOT", ""))
 
-    for key in sorted(supported_installed_versions, reverse=True):
-        ansys_version = _get_ansys_version_from_awp_root(key)
-        if ansys_version >= _EARLIEST_SUPPORTED_VERSION:
-            return supported_installed_versions[key]
     return ""
 
 
 def _get_sherlock_exe_path(year: int = None, release_number: int = None) -> str:
-    ansys_base = _get_base_ansys(year=year, release_number=release_number)
+    ansys_base, sherlock_version = _get_base_ansys(year=year, release_number=release_number)
     if not ansys_base:
         return ""
     if os.name == "nt":
         sherlock_bin = os.path.join(ansys_base, "sherlock", "SherlockClient.exe")
     else:
         sherlock_bin = os.path.join(ansys_base, "sherlock", "runSherlock")
-    return sherlock_bin
+    return sherlock_bin, sherlock_version
 
 
 def _extract_sherlock_version_year(year: int) -> int:
